% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/winratio_function.R
\name{WinRatio_sampsize}
\alias{WinRatio_sampsize}
\title{Sample Size Estimation for the Win Ratio}
\usage{
WinRatio_sampsize(
  n_arm_1,
  n_arm_2,
  alpha = 0.05,
  WinRatio = NULL,
  p = NULL,
  n.iter = 500,
  seed = 1000,
  boot = 500,
  print.iter = FALSE
)
}
\arguments{
\item{n_arm_1}{The sample size in arm 1}

\item{n_arm_2}{The sample size in arm 2}

\item{alpha}{The two-sided type I error rate}

\item{WinRatio}{The specified Win Ratio for ratio of Arm 1 wins to Arm 1 losses, either p or Win Ratio should be specified}

\item{p}{The probability of success in arm 1, either p or Win Ratio should be specified}

\item{n.iter}{The number of simulations}

\item{seed}{The seed for random number generation}

\item{boot}{The number of bootstrap samples within each simulation}

\item{print.iter}{If TRUE, each simulation number will be printed after completion}
}
\value{
A list with components including the median, 25th percentile, 75th percentile, mean, and standard deviation of Win Ratios across simulations
and the power of the proposed sample size under specified assumptions
}
\description{
A simulation-based sample size calculation approach, relying on the relationship between the win ratio and the rank distribution
}
\examples{
sim_example<-WinRatio_sampsize(n_arm_1=10,n_arm_2=10,alpha=0.05,p=0.6,n.iter=100,seed=1234,boot=100)
}
